* Add grid maker tool (it's a CLI)

+++ Model/Engine/Application.cpp
@@ -41,7 +41,7 @@ HWND CApplication::CreateGameWindow(HINSTANCE hInstance, std::wstring title, uns
 	{
 		DWORD dwStyle = GetWindowLongPtr(_hWnd, GWL_STYLE);
 		DWORD dwExStyle = GetWindowLongPtr(_hWnd, GWL_EXSTYLE);
-		RECT rect = { 0, 0, width, height };
+		RECT rect = { 0, 0, LONG(width), LONG(height) };
 		AdjustWindowRectEx(&rect, dwStyle, FALSE, dwExStyle);
 
 		ShowWindow(_hWnd, SW_SHOWNORMAL);
